Examples: 修正数据库示例，增加Int64的支持。

+++ Examples/Database/MailCSharp/Src/MySql/MySqlDataAccessManual.cs
@@ -14,7 +14,7 @@ namespace Database.MySql
 {
     public partial class MySqlDataAccess : IDataAccess
     {
-        public List<String> FromMailAttachmentSelectManyForNameById(Int Id)
+        public List<String> FromMailAttachmentSelectManyForNameById(Int64 Id)
         {
             var cmd = CreateTextCommand();
             cmd.CommandText = @"SELECT `Name` FROM `MailAttachments` WHERE `Id` = @Id";
