+++ Examples/Database/MailCSharpLinqToEntities/Src/Linq/Mail.cs
@@ -41,32 +41,32 @@ namespace Database.Linq
             return l.Single();
         }
 
-        public Mail FromMailSelectOneById(int Id)
+        public Mail FromMailSelectOneById(Int64 Id)
         {
             return dbr.Mails.ById(Id);
         }
 
-        public List<MailTo> FromMailToSelectManyById(int Id)
+        public List<MailTo> FromMailToSelectManyById(Int64 Id)
         {
             return dbr.MailTos.WhereIdIs(Id).ToList();
         }
 
-        public MailOwner FromMailOwnerSelectOneByIdAndOwnerId(int Id, int OwnerId)
+        public MailOwner FromMailOwnerSelectOneByIdAndOwnerId(Int64 Id, int OwnerId)
         {
             return dbr.MailOwners.ByIdAndOwnerId(Id, OwnerId);
         }
 
-        public int FromMailOwnerSelectCountById(int Id)
+        public int FromMailOwnerSelectCountById(Int64 Id)
         {
             return dbr.MailOwners.WhereIdIs(Id).Count();
         }
 
-        public int FromMailOwnerSelectCountByIdAndOwnerId(int Id, int OwnerId)
+        public int FromMailOwnerSelectCountByIdAndOwnerId(Int64 Id, int OwnerId)
         {
             return dbr.MailOwners.WhereIdAndOwnerIdIs(Id, OwnerId).Count();
         }
 
-        public List<MailOwner> FromMailOwnerSelectManyById(int Id)
+        public List<MailOwner> FromMailOwnerSelectManyById(Int64 Id)
         {
             return dbr.MailOwners.WhereIdIs(Id).ToList();
         }
@@ -81,12 +81,12 @@ namespace Database.Linq
             return dbr.MailOwners.WhereOwnerIdIs(OwnerId).OrderBy(mo => mo.OwnerId).ThenByDescending(mo => mo.Time).Skip(Skip).Take(Take).ToList();
         }
 
-        public List<String> FromMailAttachmentSelectManyForNameById(int Id)
+        public List<String> FromMailAttachmentSelectManyForNameById(Int64 Id)
         {
             return dbr.MailAttachments.WhereIdIs(Id).Select(ma => ma.Name).ToList();
         }
 
-        public List<MailAttachment> FromMailAttachmentSelectManyById(int Id)
+        public List<MailAttachment> FromMailAttachmentSelectManyById(Int64 Id)
         {
             return dbr.MailAttachments.WhereIdIs(Id).ToList();
         }
@@ -129,7 +129,7 @@ namespace Database.Linq
             dbr.SaveChanges();
         }
 
-        public void FromMailDeleteOneById(int Id)
+        public void FromMailDeleteOneById(Int64 Id)
         {
             foreach (var Mail in dbr.Mails.Where(m => m.Id == Id))
             {
@@ -138,7 +138,7 @@ namespace Database.Linq
             dbr.SaveChanges();
         }
 
-        public void FromMailDeleteManyToById(int Id)
+        public void FromMailDeleteManyToById(Int64 Id)
         {
             foreach (var MailTo in dbr.MailTos.Where(m => m.Id == Id))
             {
@@ -147,7 +147,7 @@ namespace Database.Linq
             dbr.SaveChanges();
         }
 
-        public void FromMailOwnerDeleteOneByIdAndOwnerId(int Id, int OwnerId)
+        public void FromMailOwnerDeleteOneByIdAndOwnerId(Int64 Id, int OwnerId)
         {
             foreach (var MailOwner in dbr.MailOwners.Where(m => m.Id == Id && m.OwnerId == OwnerId))
             {
@@ -156,7 +156,7 @@ namespace Database.Linq
             dbr.SaveChanges();
         }
 
-        public void FromMailAttachmentDeleteManyById(int Id)
+        public void FromMailAttachmentDeleteManyById(Int64 Id)
         {
             foreach (var MailAttachment in dbr.MailAttachments.Where(m => m.Id == Id))
             {
